backend/scrapper/requirements.txt
annotated-types==0.6.0
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
et-xmlfile==1.1.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gspread==5.12.3
h11==0.14.0
httplib2==0.22.0
idna==3.6
numpy==1.26.2
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.3.0.post0
pandas==2.1.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wsproto==1.2.0
xlrd==2.0.1
